Make the game work again
Quick is now Videogame (https://github.com/diogoeichert/videogame) and
with the name change the CDN link was broken.

I tried to change the game to use Videogame but it required more changes
than I wanted to do so I copied an old version of Quick that I had to
make the game work again.